drivers: opamp: delete dt-bindings/opamp/opamp.h
This was added in 26334d691b802086942560e7171a827dbfa447d8:("drivers:
Add opamp API"). It is only included by zephyr/drivers/opamp.h.

Its content needs to be in zephyr/drivers/opamp.h. The dt-bindings
directory is for sharing C macros with C and devicetree. Since this
contains an enum, which is not valid devicetree syntax, it could never
have been used except from driver code etc. Put the code where it
belongs.
